backend/pyproject.toml
[project]
name = "ircbot"
version = "0.1.0"
description = "IRC book download bot with web interface"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlalchemy>=2.0",
    "aiosqlite>=0.20",
    "irc>=20.5",
    "pyjwt>=2.9",
    "python-multipart>=0.0.18",
    "pydantic-settings>=2.7",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
